rename: timer 관련 dto 위치 변경
