+++ src/main/java/com/grepp/spring/app/model/timer/dto/response/DailyStudyLogResponse.java
@@ -1,4 +1,4 @@
-package com.grepp.spring.app.model.timer.dto;
+package com.grepp.spring.app.model.timer.dto.response;
 
 import java.time.LocalDate;
 import lombok.Getter;
